Move RunSentry to contained compilation unit w/ Init/Free API. Call Init/Free from main
